feat: adding User permission in Like endpoints

+++ app/backend/src/Controller/LikeController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Product;
+use App\Repository\UserRepository;
 use App\Service\LikeService;
 use Nelmio\ApiDocBundle\Annotation\Model;
 use OpenApi\Attributes as OA;
@@ -60,19 +61,26 @@ class LikeController extends AbstractController
             ]
         )
     )]
-    public function browse(Request $request, LikeService $likeService): JsonResponse
+    public function browse(Request $request, UserRepository $userRepository, LikeService $likeService): JsonResponse
     {
         // Parameters into query
         if ($request->query->all() !== []) {
             if (array_key_exists("user", $request->query->all())) {
 
-                $result = $likeService->findAll($request->query->all()["user"]);
+                $isUser = $userRepository->find($request->query->all()["user"]);
 
-                if ($result !== []) {
-                    return $this->json($result, Response::HTTP_OK, [], ["groups" => ["read:Product:item"]]);
-                } else {
-                    return $this->json(["code" => 500, "message" => "Internal Server Error"], Response::HTTP_NOT_FOUND);
+                if ($this->isGranted('user_read', $isUser)) {
+
+                    $result = $likeService->findAll($request->query->all()["user"]);
+
+                    if ($result !== []) {
+                        return $this->json($result, Response::HTTP_OK, [], ["groups" => ["read:Product:item"]]);
+                    }
+
+                    return $this->json(["code" => 404, "message" => "No Likes was found"], Response::HTTP_NOT_FOUND);
                 }
+
+                return $this->json(["code" => 403, "message" => "Forbidden"], Response::HTTP_FORBIDDEN);
             }
         }
 
@@ -137,17 +145,25 @@ class LikeController extends AbstractController
             ]
         )
     )]
-    public function add(Request $request, LikeService $likeService): JsonResponse
+    public function add(Request $request, UserRepository $userRepository, LikeService $likeService): JsonResponse
     {
         // Parameters into query
         if ($request->query->all() !== []) {
             if (array_key_exists("product", $request->query->all()) && array_key_exists("user", $request->query->all())) {
 
-                $result = $likeService->add($request->query->all()["product"], $request->query->all()["user"]);
+                $isUser = $userRepository->find($request->query->all()["user"]);
 
-                if ($result) {
-                    return $this->json(null, Response::HTTP_NO_CONTENT, []);
+                if ($this->isGranted("user_add", $isUser)) {
+                    $result = $likeService->add($request->query->all()["product"], $request->query->all()["user"]);
+
+                    if ($result) {
+                        return $this->json(null, Response::HTTP_NO_CONTENT, []);
+                    }
+
+                    return $this->json(["code" => 404, "message" => "No Likes was found"], Response::HTTP_NOT_FOUND);
                 }
+
+                return $this->json(["code" => 403, "message" => "Forbidden"], Response::HTTP_FORBIDDEN);
             }
         }
 
@@ -212,17 +228,25 @@ class LikeController extends AbstractController
             ]
         )
     )]
-    public function delete(Request $request, LikeService $likeService): JsonResponse
+    public function delete(Request $request, UserRepository $userRepository, LikeService $likeService): JsonResponse
     {
         // Parameters into query
         if ($request->query->all() !== []) {
             if (array_key_exists("product", $request->query->all()) && array_key_exists("user", $request->query->all())) {
 
-                $result = $likeService->delete($request->query->all()["product"], $request->query->all()["user"]);
+                $isUser = $userRepository->find($request->query->all()["user"]);
 
-                if ($result) {
-                    return $this->json(null, Response::HTTP_NO_CONTENT, []);
+                if ($this->isGranted("user_delete", $isUser)) {
+                    $result = $likeService->delete($request->query->all()["product"], $request->query->all()["user"]);
+
+                    if ($result) {
+                        return $this->json(null, Response::HTTP_NO_CONTENT, []);
+                    }
+
+                    return $this->json(["code" => 404, "message" => "No Likes was found"], Response::HTTP_NOT_FOUND);
                 }
+
+                return $this->json(["code" => 403, "message" => "Forbidden"], Response::HTTP_FORBIDDEN);
             }
         }
